Share OpenStack provider test fixtures
Move the fake OpenStack server and Kubernetes client setup out of the
provider race test and into shared package test helpers.

Reuse one service catalog across VolumeClass cache and provider race
tests to prevent their OpenStack bootstrap fixtures from diverging.
